requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
httpx>=0.25.2
python-dotenv>=1.0.0
pydantic>=2.8.0
pydantic-settings>=2.1.0

fastapi
httpx
pydantic
pydantic-settings
python-dotenv
uvicorn[standard]
